requirements.txt
sqlalchemy
pandas
ccxt
pyyaml
coloredlogs
